Moved helpers.js to the root of a components folder.
