+++ src/components/Dataset/Details/Form.js
@@ -8,7 +8,7 @@ import { search as searchOrganizations } from '../../../api/organization';
 import { search as searchInstallations } from '../../../api/installation';
 import { searchDatasets } from '../../../api/dataset';
 import { getDatasetSubtypes, getDatasetTypes, getMaintenanceUpdateFrequencies } from '../../../api/enumeration';
-import { getPermittedOrganizations, prettifyLicense } from '../../../api/util/helpers';
+import { getPermittedOrganizations, prettifyLicense } from '../../helpers';
 import { FilteredSelectControl } from '../../widgets';
 import formValidationWrapper from '../../hoc/formValidationWrapper';
 import withContext from '../../hoc/withContext';
