+++ src/components/Dataset/index.js
@@ -26,7 +26,7 @@ import { ContactList, EndpointList, IdentifierList, TagList, MachineTagList, Com
 import ConstituentsDataset from './ConstituentsDataset';
 import MenuConfig from './menu.config';
 import withContext from '../hoc/withContext';
-import { getSubMenu } from '../../api/util/helpers';
+import { getSubMenu } from '../helpers';
 import AuthRoute from '../AuthRoute';
 import PermissionWrapper from '../hoc/PermissionWrapper';
