+++ src/components/Institution/index.js
@@ -19,7 +19,7 @@ import { ContactList, IdentifierList, TagList } from '../common';
 import Exception404 from '../exception/404';
 import MenuConfig from './menu.config';
 import withContext from '../hoc/withContext';
-import { getSubMenu } from '../../api/util/helpers';
+import { getSubMenu } from '../helpers';
 
 class Institution extends Component {
   constructor(props) {
