+++ src/components/Node/index.js
@@ -14,7 +14,7 @@ import Installations from './Installations';
 import Exception404 from '../exception/404';
 import MenuConfig from './menu.config';
 import withContext from '../hoc/withContext';
-import { getSubMenu } from '../../api/util/helpers';
+import { getSubMenu } from '../helpers';
 import AuthRoute from '../AuthRoute';
 
 class NodeItem extends Component {
