+++ src/components/Organization/index.js
@@ -28,7 +28,7 @@ import Installations from './Installations';
 import Exception404 from '../exception/404';
 import MenuConfig from './menu.config';
 import withContext from '../hoc/withContext';
-import { getSubMenu } from '../../api/util/helpers';
+import { getSubMenu } from '../helpers';
 import AuthRoute from '../AuthRoute';
 
 class Organization extends Component {
